+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.6 [unreleased]
 
+- Added `update-policy` command
 - 2x performance for object commands
 
 ## 0.1.5